changed form uns to long

+++ 0x0A-argc_argv/4-add.c
@@ -12,7 +12,7 @@
 
 int main(int argc, char *argv[])
 {
-	unsigned int aux = 0, suma = 0;
+	long aux = 0, suma = 0;
 	int i;
 
 	for (i = 1; i < argc; i++)
